feat: split map dialog into mobile and desktop component
