+++ package.json
@@ -23,6 +23,7 @@
     "@radix-ui/react-visually-hidden": "1.1.1",
     "@tabler/icons-react": "3.26.0",
     "@types/lodash": "4.17.13",
+    "@uidotdev/usehooks": "2.4.1",
     "axios-retry": "4.5.0",
     "class-variance-authority": "0.7.1",
     "immer": "10.1.1",
